Fix params in helper

+++ helpers.go
@@ -10,4 +10,4 @@ func Bool(v bool) *bool { return &v }
 func Int(v int) *int { return &v }
 
 // Float64 helper allows to quickly get a pointer to a float46
-func Float64(v int) *int { return &v }
+func Float64(v float64) *float64 { return &v }
